修复导入路径错误和组织通讯录数据加载问题
- 修复 NoteList.vue 导入 types.js -> types
- 修复 useNoteListContextMenu.ts 导入 .ts 扩展名
- 修复 organize_dept 表 is_deleted 字段映射
- 更新 organize_position 表字段映射 (post_code <- ID)
- 修复 organize_dept 模型添加 udm_org_id, udm_org_parent_id 字段

+++ backend/internal/repository/model/organize_dept.go
@@ -7,7 +7,7 @@ type OrganizeDept struct {
 	ParentId  int       `gorm:"column:parent_id;" json:"parent_id"`                       // 父部门id
 	Ancestors string    `gorm:"column:ancestors;" json:"ancestors"`                       // 祖级列表
 	DeptName  string    `gorm:"column:dept_name;" json:"dept_name"`                       // 部门名称
-	OrderNum  int       `gorm:"column:order_num;" json:"order_num"`                       // 显示顺序
+	OrderNum  float64   `gorm:"column:order_num;" json:"order_num"`                       // 显示顺序
 	Leader    string    `gorm:"column:leader;" json:"leader"`                             // 负责人
 	Phone     string    `gorm:"column:phone;" json:"phone"`                               // 联系电话
 	Email     string    `gorm:"column:email;" json:"email"`                               // 邮箱
